server/requirements.txt
alembic==1.5.7
aniso8601==7.0.0
appnope==0.1.2
autopep8==1.5.5
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.7.1
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-GraphQL==2.0.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
folium==0.12.1
GeoAlchemy2==0.8.4
graphene==2.1.8
graphene-file-upload==1.3.0
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.12.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.0.0
html5lib==1.0.1
idna==2.10
importlib-metadata==4.0.1
infomap==1.3.0
infostop==0.1.8
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.0.1
lxml==4.3.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.0.2
numexpr==2.6.9
numpy==1.20.2
olefile==0.46
opencv-python==4.5.1.48
pandas==1.2.4
parso==0.8.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.17
psycopg2==2.8.6
ptyprocess==0.7.0
pybind11==2.6.2
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.2.0
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
Rx==1.6.1
scipy==1.6.2
Shapely==1.7.1
singledispatch==3.6.1
six==1.15.0
soupsieve==1.8
SQLAlchemy==1.4.6
SQLAlchemy-Utils==0.37.4
tables==3.4.4
termcolor==1.1.0
toml==0.10.2
traitlets==5.0.5
twilio==6.56.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
